Dedupe `lib` `cutscene` for PSP
