+++ config/splat.pspeu.stlib.yaml
@@ -73,6 +73,9 @@ segments:
       - [0x1D498, c, lib_psp/1D498]
       - [0x1D650, c, lib/e_chair]
       - [0x1D9E8, c, lib_psp/unk_1D9E8]
+      - [0x231B0, c, lib_psp/lang]
+      - [0x23250, c, lib_psp/cutscene]
+      - [0x25CD0, c, lib_psp/unk_25CD0]
       - [0x290C8, c, lib_psp/e_shop]
       - [0x2BC10, c, lib_psp/2BC10]
       - [0x36780, data]
@@ -97,6 +100,9 @@ segments:
       - [0x38C78, data]
       - [0x60AB8, .data, lib/e_chair]
       - [0x60AE8, data]
+      - [0x60E50, .data, lib_psp/cutscene]
+      - [0x61158, data, cutscene] # LIB_EntityCutscene data
+      - [0x61218, data]
       - [0x6D800, .rodata, lib_psp/e_spellbook_magic_tome]
       - [0x6D858, .rodata, lib/e_life_up]
       - [0x6D878, .rodata, lib_psp/e_dhuron]
@@ -108,6 +114,8 @@ segments:
       - [0x6D988, .rodata, lib/en_thornweed_corpseweed]
       - [0x6D9C8, .rodata, lib_psp/1B7F0]
       - [0x6D9E8, .rodata, lib_psp/unk_1D9E8]
+      - [0x6DA20, .rodata, lib_psp/cutscene]
+      - [0x6DAD8, .rodata, lib_psp/unk_25CD0]
       - [0x6DB78, .rodata, lib_psp/e_shop]
       - [0x6DBB8, .rodata, lib_psp/2BC10]
       - [0x6DD80, .bss, lib/create_entity]
