+++ config/symbols.pspeu.stlib.txt
@@ -105,6 +105,16 @@ EntityCorpseweedProjectile = 0x09252858;
 EntityBreakable = 0x092546D0;
 EntityChair = 0x09254DC0;
 func_us_801BB8DC = 0x09255D78;
+GetLang = 0x0925A830;
+DrawCutsceneActorName = 0x0925A8D0;
+CutsceneUnk3 = 0x0925ABE8;
+CutsceneUnk4 = 0x0925AC70;
+CutsceneUnk1 = 0x0925AD40;
+SetCutsceneScript = 0x0925ADA0;
+SetCutsceneEnd = 0x0925AFB0;
+CutsceneRun = 0x0925AFE8;
+CutsceneSkip = 0x0925B380;
+LIB_EntityCutscene = 0x0925B470;
 EntityLibrarianChair = 0x0925E018;
 explosionVariantSizes = 0x0926F738;
 g_UnkRecursPrim2Inds = 0x0926F798;
@@ -158,8 +168,10 @@ obtainedStr = 0x092A5468;
 D_psp_0924BC68 = 0x092A5470;
 D_psp_0924BC70 = 0x092A5478;
 D_psp_0924BC78 = 0x092A5480;
+LIB_cutscene_data = 0x092A54F8;
 D_us_80181ACC = 0x092A5700;
 g_IsCutsceneDone = 0x092A5728;
-g_Dialogue = 0x092A5CDC; // size:0x44
+cutsceneUnk3Unknown = 0x092A5938;
+g_Dialogue = 0x092A5CE0; // size:0x40
 g_SkipCutscene = 0x092A5D20;
 g_CutsceneFlags = 0x092A5D28;
