+++ config/symbols.pspeu.txt
@@ -57,6 +57,7 @@ g_pStObjLayoutVertical = 0x08C6BC34;
 g_pStObjLayoutHorizontal = 0x08C6BC38;
 g_PlOvl = 0x08CE9C40;
 DestroyEntitiesFromIndex = 0x0913E838;
+g_CutsceneHasControl = 0x091CE598;
 g_StageId = 0x091CE5B8;
 g_IsUsingCd = 0x091CE5C8;
 g_LoadFile = 0x091CE5D8;
