+++ src/st/cutscene_unk1.h
@@ -3,11 +3,15 @@
 #define CUTSCENE_UNK1_NEXT_X 2
 #endif
 
+#ifndef CUTSCENE_UNK1_UNK17
+#define CUTSCENE_UNK1_UNK17 8
+#endif
+
 static void CutsceneUnk1(void) {
     g_Dialogue.nextLineY = g_Dialogue.startY + 0x14;
     g_Dialogue.nextCharX = g_Dialogue.nextLineX = CUTSCENE_UNK1_NEXT_X;
     g_Dialogue.nextCharY = 0;
     g_Dialogue.unk12 = 0;
     g_Dialogue.nextCharTimer = 0;
-    g_Dialogue.unk17 = 8;
+    g_Dialogue.unk17 = CUTSCENE_UNK1_UNK17;
 }
