+++ src/st/cutscene_unk3.h
@@ -4,17 +4,17 @@
 #endif
 
 #if defined(VERSION_PSP)
-extern void* D_pspeu_0927B6A8;
+extern s32 cutsceneUnk3Unknown[0x380]; // bss
 #endif
 
-static void CutsceneUnk3(s16 arg0) {
+static void CutsceneUnk3(s16 nextCharY) {
     RECT rect;
     rect.x = CUTSCENE_UNK3_RECT_X;
-    rect.y = (arg0 * 12) + 384;
+    rect.y = (nextCharY * 12) + 384;
     rect.w = 64;
     rect.h = 12;
     ClearImage(&rect, 0, 0, 0);
 #if defined(VERSION_PSP)
-    func_890A3C0(&D_pspeu_0927B6A8, 0, 0x380);
+    func_890A3C0(cutsceneUnk3Unknown, 0, LEN(cutsceneUnk3Unknown));
 #endif
 }
