+++ src/st/lib/cutscene.c
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "lib.h"
-#include "../pfn_entity_update.h"
 #include "../cutscene.h"
 
 extern u8 D_us_80183F60;
