+++ src/st/lib_psp/lang.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../lib/lib.h"
+
+#include "../get_lang.h"
